# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: AI Enhanced PDF Scholar API
  description: |
    **Enterprise-grade intelligent document management platform for academic research**
    
    This API provides comprehensive functionality for:
    - 🎯 **Intelligent Document Processing**: Smart import with metadata extraction and duplicate detection
    - 📊 **Advanced Citation Analysis**: Multi-format parsing with network discovery
    - 🧠 **RAG (Retrieval-Augmented Generation)**: AI-powered document queries and analysis
    - 🔐 **Enterprise Security**: JWT authentication with RBAC and XSS protection
    - ⚡ **Real-time Features**: WebSocket support for streaming operations
    - 📈 **Performance Monitoring**: APM with metrics and health checks
    
    ## Authentication
    
    Most endpoints require JWT authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    
    API endpoints are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute per IP
    - **Global IP**: 1000 requests per hour per IP
    - **Authenticated users**: Higher limits apply
    
    ## Error Handling
    
    The API uses standard HTTP status codes and provides detailed error information:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource not found)
...
    
  version: "2.1.0"
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [authentication]
      summary: User login
      description: Authenticate user and receive JWT access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: Username or email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  description: User password
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/auth/refresh:
    post:
      tags: [authentication]
      summary: Refresh JWT token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Document Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            error_code: "UNAUTHORIZED"
...
  schemas:
...
    BaseResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error_code:
              type: string
              description: Machine-readable error code
            details:
              type: object
              description: Additional error details
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                  description: JWT access token
                refresh_token:
                  type: string
                  description: JWT refresh token
                token_type:
                  type: string
                  example: "bearer"
                expires_in:
                  type: integer
                  description: Token expiration time in seconds
                user:
                  $ref: '#/components/schemas/UserProfile'
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                expires_in:
                  type: integer
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, researcher]
        permissions:
          type: array
          items:
            type: string

    # Document Models
...